LatticeGeometry

Documentation for LatticeGeometry.

LatticeGeometry.AbstractCrystalGeometer

LatticeGeometry.AbstractionError

LatticeGeometry.DefaultCrystalGeometer

LatticeGeometry.UnitCell

LatticeGeometry.UnitCell

LatticeGeometry._positions_within_bounds

LatticeGeometry.construct_metric

LatticeGeometry.AbstractCrystalGeometer — Type

abstract type AbstractCrystalGeometer end

A generic family of structures/functors that will control the access to metrics and distances based on basis vectors.

LatticeGeometry.AbstractionError — Type

AbstractionError <: Exception

A pretty generic catch all for abtract typing on the fly.

LatticeGeometry.DefaultCrystalGeometer — Type

DefaultCrystalGeometer{D, T}

The go-to AbstractCrystalGeometer one can use for systems of spatial dimension D and with basis vectors of with eltype T <: AbstractFloat.

LatticeGeometry.UnitCell — Type

UnitCell{Natoms, Dimension, Type} <: AbstractUnitCell

Generic UnitCell structure with Natoms in a given spatial Dimension. Type signifies the numerical type used to store the positions. We use the standard convention of writing the positions in the cell coordinates, where each component of the positions must lie on [0,1].

LatticeGeometry.UnitCell — Method

UnitCell([atom_names], atomic_positions)

Constructor for the UnitCell. The atomic_positions must be supplied as a StaticArrays.SVector of StaticArrays.SVectors to optimize memory. The atom_names are optional, though if they are supplied then length(atom_names) == length(atomic_positions). If they are not supplied, default atom_names will be used.

julia> using StaticArrays

julia> positions = @SVector [ SVector(0.0, 0.0, 0.0), SVector(0.5, 0.5, 0.5) ];

julia> labels = SVector( "Fe", "As" );

julia> UnitCell(labels, positions)

  `UnitCell` containing 2 atoms in 3 spatial dimensions.

    Fe at [0.0, 0.0, 0.0]
    As at [0.5, 0.5, 0.5]

julia> UnitCell(positions; warn_on = false)

  `UnitCell` containing 2 atoms in 3 spatial dimensions.

    Atom_1 at [0.0, 0.0, 0.0]
    Atom_2 at [0.5, 0.5, 0.5]

LatticeGeometry._positions_within_bounds — Method

_positions_within_bounds(positions; [interval = [0, 1]])

Internal function to check that each position coordinate is on the supplied interval.

LatticeGeometry.construct_metric — Method

construct_metric(lattice_vectors)

Construct the crystallographic metric tensor $g_{ij}$

This metric tensor maps distances in the mathematical ReducedBasis to distances in the physical CrystalBasis using the lattice_vectors.

Let $\vec{P}$ and $\vec{Q}$ be points in space. In the CrystalBasis, these vectors are parameterized by

\[\begin{aligned} \vec{P} &= \sum_{i=1}^d p_i \vec{a}_i, \\ \vec{Q} &= \sum_{i=1}^d q_i \vec{a}_i. \end{aligned}\]

The collection of coefficients $\{p_i\}$ and $\{q_i\}$ are the ReducedBasis coordinates. We denote these vectors with boldface font and write them as

\[\begin{aligned} \boldsymbol{P} &= \sum_{i=1}^d p_i \hat{e}_i, \\ \boldsymbol{Q} &= \sum_{i=1}^d q_i \hat{e}_i. \end{aligned}\]

Then the scalar product has

\[\vec{P}\cdot\vec{Q} = \sum_{i,j} p_iq_j \vec{a}_i\cdot\vec{a}_j \equiv \boldsymbol{P} \cdot \mathsf{g} \cdot \boldsymbol{Q},\]

where the metric tensor $\mathsf{g}$ has the components:

\[g_{ij} = \vec{a}_i \cdot \vec{a}_j.\]