LatticeGeometry

Documentation for LatticeGeometry.

LatticeGeometry.AbstractionError

LatticeGeometry.UnitCell

LatticeGeometry.UnitCell

LatticeGeometry._positions_within_bounds

LatticeGeometry.AbstractionError — Type

AbstractionError <: Exception

A pretty generic catch all for abtract typing on the fly.

LatticeGeometry.UnitCell — Type

UnitCell{Natoms, Dimension, Type} <: AbstractUnitCell

Generic UnitCell structure with Natoms in a given spatial Dimension. Type signifies the numerical type used to store the positions. We use the standard convention of writing the positions in the cell coordinates, where each component of the positions must lie on [0,1].

LatticeGeometry.UnitCell — Method

UnitCell([atom_names], atomic_positions)

Constructor for the UnitCell. The atomic_positions must be supplied as a StaticArrays.SVector of StaticArrays.SVectors to optimize memory. The atom_names are optional, though if they are supplied then length(atom_names) == length(atomic_positions). If they are not supplied, default atom_names will be used.

julia> using StaticArrays

julia> positions = @SVector [ SVector(0.0, 0.0, 0.0), SVector(0.5, 0.5, 0.5) ];

julia> labels = SVector( "Fe", "As" );

julia> UnitCell(labels, positions)

  `UnitCell` containing 2 atoms in 3 spatial dimensions.

    Fe at [0.0, 0.0, 0.0]
    As at [0.5, 0.5, 0.5]

julia> UnitCell(positions; warn_on = false)

  `UnitCell` containing 2 atoms in 3 spatial dimensions.

    Atom_1 at [0.0, 0.0, 0.0]
    Atom_2 at [0.5, 0.5, 0.5]

LatticeGeometry._positions_within_bounds — Method

_positions_within_bounds(positions; [interval = [0, 1]])

Internal function to check that each position coordinate is on the supplied interval.